Now without that BaseHTTPServer nonsense
which was cribbed from fbconsole's implementation. No reason to use
that; we already have the Django webserver running! This re-implements
that flow in a more Django-friendly way, using standard URL routing and
redirect() to bounce the user around the OAuth flow.

This uses sqlite3-backed sessions to store intermediate state.