Time series loading now works
with d3.json(), which is used to grab ~3M of data from /get-user-data
and marshal it into useful form.

It might be worthwhile to precompute these time series, actually, since
the frontend computation currently ties up the browser for ~5-10
seconds (!)

Also changed /sync-user-data to grab only up to the last date where data
was synced, so that we avoid populating foo_userdata table with useless
zeros. Added Fitbit.get_user_last_sync_date() to support this.